requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.0
sqlite3==2.6.0
pillow==9.0.0